Update index.css
Hi Bleda. I did the changes you sugested unless the on you refer to the linear-gradient. I do think it matters for the fade the pic does to the black background, anyway. I did the changes on github web because I dont have conction with anymore with this old repositoris.